ref: webhook handling to be less magic
Replace the type annotation introspection with an explicit if statement.
