--- kodiak/github/README.md
@@ -1,4 +1,0 @@
-# github-webhooks
-A library to build typesafe Github webhook handlers
-
-WIP
