--- kodiak/github/fixtures/hook_configuration.json
@@ -1,18 +1,0 @@
-{
-    "id": 1,
-    "url": "https://api.github.com/repos/octocat/Hello-World/hooks/1",
-    "test_url": "https://api.github.com/repos/octocat/Hello-World/hooks/1/test",
-    "ping_url": "https://api.github.com/repos/octocat/Hello-World/hooks/1/pings",
-    "name": "web",
-    "events": [
-        "push",
-        "pull_request"
-    ],
-    "active": true,
-    "config": {
-        "url": "http://example.com/webhook",
-        "content_type": "json"
-    },
-    "updated_at": "2011-09-06T20:39:23Z",
-    "created_at": "2011-09-06T17:26:27Z"
-}
