+++ kodiak/test_main.py
@@ -1,6 +1,16 @@
+import json
+from pathlib import Path
+from typing import Tuple
+
+import pytest
+from pytest_mock import MockFixture
+from starlette import status
 from starlette.testclient import TestClient
 
-from kodiak.main import get_branch_name
+from kodiak import app_config as conf
+from kodiak.main import app
+from kodiak.test_events import MAPPING
+from kodiak.test_utils import wrap_future
 
 
 def test_read_main(client: TestClient) -> None:
@@ -9,9 +19,83 @@ def test_read_main(client: TestClient) -> None:
     assert response.json() == "OK"
 
 
-def test_get_branch_name() -> None:
-    assert get_branch_name("refs/heads/master") == "master"
-    assert (
-        get_branch_name("refs/heads/master/refs/heads/123") == "master/refs/heads/123"
+@pytest.fixture
+def client() -> TestClient:
+    return TestClient(app)
+
+
+def get_body_and_hash(data: dict) -> Tuple[bytes, str]:
+    import hmac
+    import hashlib
+
+    import ujson
+
+    body = ujson.dumps(data).encode()
+    sha = hmac.new(
+        key=conf.SECRET_KEY.encode(), msg=body, digestmod=hashlib.sha1
+    ).hexdigest()
+    return body, sha
+
+
+@pytest.mark.parametrize("event_name", (event_name for event_name, _schema in MAPPING))
+def test_event_parsing(
+    client: TestClient, event_name: str, mocker: MockFixture
+) -> None:
+    """Test all of the events we have"""
+    handle_webhook_event = mocker.patch(
+        "kodiak.main.handle_webhook_event", return_value=wrap_future(None)
     )
-    assert get_branch_name("refs/tags/v0.1.0") is None
+    for index, fixture_path in enumerate(
+        (Path(__file__).parent / "test" / "fixtures" / "events" / event_name).rglob(
+            "*.json"
+        )
+    ):
+        data = json.loads(fixture_path.read_bytes())
+
+        body, sha = get_body_and_hash(data)
+
+        assert handle_webhook_event.call_count == index
+        res = client.post(
+            "/api/github/hook",
+            data=body,
+            headers={"X-Github-Event": event_name, "X-Hub-Signature": sha},
+        )
+        assert res.status_code == status.HTTP_200_OK
+        assert handle_webhook_event.call_count == index + 1
+
+
+def test_event_parsing_missing_github_event(
+    client: TestClient, mocker: MockFixture
+) -> None:
+    handle_webhook_event = mocker.patch(
+        "kodiak.main.handle_webhook_event", return_value=wrap_future(None)
+    )
+    data = {"hello": 123}
+
+    body, sha = get_body_and_hash(data)
+
+    assert handle_webhook_event.called is False
+    res = client.post("/api/github/hook", data=body, headers={"X-Hub-Signature": sha})
+    assert res.status_code == status.HTTP_400_BAD_REQUEST
+    assert handle_webhook_event.called is False
+
+
+def test_event_parsing_invalid_signature(
+    client: TestClient, mocker: MockFixture
+) -> None:
+    handle_webhook_event = mocker.patch(
+        "kodiak.main.handle_webhook_event", return_value=wrap_future(None)
+    )
+    data = {"hello": 123}
+
+    # use a different dict for the signature so we get an signature mismatch
+    _, sha = get_body_and_hash({})
+
+    assert handle_webhook_event.called is False
+    res = client.post(
+        "/api/github/hook",
+        json=data,
+        headers={"X-Github-Event": "content_reference", "X-Hub-Signature": sha},
+    )
+    assert res.status_code == status.HTTP_400_BAD_REQUEST
+    assert handle_webhook_event.called is False
